Redo logo as SVG to work around bg clipping gradient bug
PROBLEM: The Infrastructure Stuff logo is currently implemented in pure
CSS and appears throughout the site, but the background gradient
disappears as you navigate around.

SOLUTION: We're using SVG instead.

Part of being a good tactician at war is knowing when to retreat. Early
victories in the foreign lands of "web standards" and "Next.js" were
encouraging, but the season has turned and now the tanks are stuck in
the mud. Troops (i.e., I) complain about being poorly-equipped for the
attritive war against the sprawling plane of knowledge known as the
"Mozilla Developer Network" and its borderline-DDoS-against-your-brain
comprehensiveness. The temptation to turn and flee to a familiar refuge
is too great.

Thus ends our attempt to stand on the bleeding edge of web standards.